+++ apps/desktop/AStar.Dev.OneDrive.Sync.Client.Tests.Unit/Settings/GivenASettingsViewModel.cs
@@ -1,4 +1,3 @@
-using AStar.Dev.OneDrive.Sync.Client.Conflicts;
 using AStar.Dev.OneDrive.Sync.Client.Data.Repositories;
 using AStar.Dev.OneDrive.Sync.Client.Domain;
 using AStar.Dev.OneDrive.Sync.Client.Infrastructure.Shell;
