+++ apps/desktop/AStar.Dev.OneDrive.Sync.Client.Tests.Unit/Splash/GivenASplashWindowViewModel.cs
@@ -5,7 +5,7 @@ public sealed class GivenASplashWindowViewModel
     [Fact]
     public void when_created_then_status_is_empty()
     {
-        var sut = new AStar.Dev.OneDrive.Sync.Client.Splash.SplashWindowViewModel();
+        var sut = new Client.Splash.SplashWindowViewModel();
 
         sut.Status.ShouldBeEmpty();
     }
@@ -13,7 +13,7 @@ public sealed class GivenASplashWindowViewModel
     [Fact]
     public void when_status_is_set_then_property_changed_is_raised()
     {
-        var sut = new AStar.Dev.OneDrive.Sync.Client.Splash.SplashWindowViewModel();
+        var sut = new Client.Splash.SplashWindowViewModel();
         var raised = false;
         sut.PropertyChanged += (_, args) =>
         {
@@ -29,7 +29,7 @@ public sealed class GivenASplashWindowViewModel
     [Fact]
     public void when_status_is_set_then_new_value_is_returned()
     {
-        var sut = new AStar.Dev.OneDrive.Sync.Client.Splash.SplashWindowViewModel();
+        var sut = new Client.Splash.SplashWindowViewModel();
 
         sut.Status = "Configuring database…";
