+++ apps/desktop/AStar.Dev.OneDrive.Sync.Client.Tests.Unit/Startup/GivenShellServiceExtensions.cs
@@ -1,6 +1,5 @@
 using System.IO.Abstractions;
 using AStar.Dev.OneDrive.Sync.Client.Data;
-using AStar.Dev.OneDrive.Sync.Client.Infrastructure;
 using AStar.Dev.OneDrive.Sync.Client.Infrastructure.ApplicationConfiguration;
 using AStar.Dev.OneDrive.Sync.Client.LogViewer;
 using AStar.Dev.OneDrive.Sync.Client.Startup;
